First initial commit to the new master branch.
This is the first commit before the release it contains all relevant changes
happened during the project. We create this branch to reset the project log.
The old master branch it at the archives/old-flow.

The new master branch is managed with Lerna; this is a multi-package
Repository manager. With his tool we can efficiently manage versioning
packages that are co-dependent; still, you want to share useful
functionality with others, so the best of the conquering techniques.
You don't have to compromise if you want your monolithic system
operation in a smart way (cathedral-style development) and yet
be able the share chunks of finely grained functionalities to the masses.